Split presence into separate file
